Fix graphviz-output of CoBüchiAutomaton and SafetyAutomaton
The graphviz-rendering of Automata used the label of the state as the id of the corresponding node in graphviz. This would fail if the state label is not a valid graphviz id, e.g., if it contains punctuation marks.
By enclosing the label in quotation marks, this is circumvented.

Moreover, the graphviz-rendering of safety automata used the hashvalue of the states instead of their label as the id, leading to graphviz-representations that change between two executions of BoSy.
We now use the label of the state in safety automata as the node label in graphviz, analogously to the graphviz-rendering of CoBüchiAutomata.